Course Level Progress Working

+++ GroupDjango/enrolment/models.py
@@ -6,7 +6,7 @@ class Enrolment(models.Model):
     user = models.ForeignKey('accounts.CustomUser', on_delete= models.CASCADE)
     course = models.ForeignKey('courses.Course', on_delete= models.CASCADE)
     access = models.BooleanField(default=True)
-    grade = models.PositiveSmallIntegerField(null=True, blank=True)
+    grade = models.FloatField(null=True, blank=True)
 
     def __str__(self):
         str = self.user.username + "-" + self.course.name
